python_code/python_GUI/requirements.txt
numpy==1.19.1
PyQt5==5.15.0
pyqtgraph==0.11.0
pyserial==3.4
QDarkStyle==2.8.1
QLed==1.3.1
matplotlib==3.2.1